+++ tests/integration/test_locked_resolve.py
@@ -55,9 +55,7 @@ def create_lock_observer(lock_configuration):
         resolver=ConfiguredResolver(pip_configuration=pip_configuration),
         wheel_builder=WheelBuilder(
             package_index_configuration,
-            prefer_older_binary=pip_configuration.prefer_older_binary,
-            use_pep517=pip_configuration.use_pep517,
-            build_isolation=pip_configuration.build_isolation,
+            build_configuration=pip_configuration.build_configuration,
         ),
         package_index_configuration=package_index_configuration,
     )